refactor types and configs
