+++ packages/reva-app/src/interface.ts
@@ -1,3 +1,7 @@
+export type Page = "show-results" | "load-submission";
+export type Direction = "previous" | "next";
+export type Navigation = { direction: Direction; page: Page };
+
 export interface Certificate {
   id: string;
   description: string;
